Initial changes to allow loging into local users

+++ src/ssh/azext_ssh/ssh_utils.py
@@ -19,47 +19,48 @@ CLEANUP_TIME_INTERVAL_IN_SECONDS = 10
 CLEANUP_AWAIT_TERMINATION_IN_SECONDS = 30
 
 
-def start_ssh_connection(port, ssh_args, ip, username, cert_file, private_key_file, delete_keys):
+def start_ssh_connection(port, ssh_args, ip, username, cert_file, private_key_file, delete_keys, delete_cert):
 
     ssh_arg_list = []
     if ssh_args:
         ssh_arg_list = ssh_args
-
+    
     log_file = None
-    if '-E' not in ssh_arg_list and set(['-v', '-vv', '-vvv']).isdisjoint(ssh_arg_list):
-        # If the user either provides his own client log file (-E) or
-        # wants the client log messages to be printed to the console (-vvv/-vv/-v),
-        # we should not use the log files to check for connection success.
-        log_file_dir = os.path.dirname(cert_file)
-        log_file_name = 'ssh_client_log_' + str(os.getpid())
-        log_file = os.path.join(log_file_dir, log_file_name)
-        ssh_arg_list = ssh_arg_list + ['-E', log_file, '-v']
+    if delete_keys or delete_cert:
+        if '-E' not in ssh_arg_list and set(['-v', '-vv', '-vvv']).isdisjoint(ssh_arg_list):
+            # If the user either provides his own client log file (-E) or
+            # wants the client log messages to be printed to the console (-vvv/-vv/-v),
+            # we should not use the log files to check for connection success.
+            log_file_dir = os.path.dirname(cert_file)
+            log_file_name = 'ssh_client_log_' + str(os.getpid())
+            log_file = os.path.join(log_file_dir, log_file_name)
+            ssh_arg_list = ssh_arg_list + ['-E', log_file, '-v']
+        # Create a new process that will wait until the connection is established and then delete keys.
+        cleanup_process = mp.Process(target=_do_cleanup, args=(delete_keys, delete_cert, cert_file, private_key_file,
+                                    log_file, True))
+        cleanup_process.start()
 
     command = [_get_ssh_path(), _get_host(username, ip)]
     command = command + _build_args(cert_file, private_key_file, port) + ssh_arg_list
 
-    # Create a new process that will wait until the connection is established and then delete keys.
-    cleanup_process = mp.Process(target=_do_cleanup, args=(delete_keys, cert_file, private_key_file,
-                                 log_file, True))
-    cleanup_process.start()
-
     logger.debug("Running ssh command %s", ' '.join(command))
     subprocess.call(command, shell=platform.system() == 'Windows')
 
-    if cleanup_process.is_alive():
-        cleanup_process.terminate()
-        # wait for process to terminate
-        t0 = time.time()
-        while cleanup_process.is_alive() and (time.time() - t0) < CLEANUP_AWAIT_TERMINATION_IN_SECONDS:
-            time.sleep(1)
+    if delete_keys or delete_cert:
+        if cleanup_process.is_alive():
+            cleanup_process.terminate()
+            # wait for process to terminate
+            t0 = time.time()
+            while cleanup_process.is_alive() and (time.time() - t0) < CLEANUP_AWAIT_TERMINATION_IN_SECONDS:
+                time.sleep(1)
 
-    # Make sure all files have been properly removed.
-    _do_cleanup(delete_keys, cert_file, private_key_file)
-    if log_file:
-        file_utils.delete_file(log_file, f"Couldn't delete temporary log file {cert_file}. ", True)
-    if delete_keys:
-        temp_dir = os.path.dirname(cert_file)
-        file_utils.delete_folder(temp_dir, f"Couldn't delete temporary folder {temp_dir}", True)
+        # Make sure all files have been properly removed.
+        _do_cleanup(delete_keys, cert_file, private_key_file)
+        if log_file:
+            file_utils.delete_file(log_file, f"Couldn't delete temporary log file {cert_file}. ", True)
+        if delete_keys:
+            temp_dir = os.path.dirname(cert_file)
+            file_utils.delete_folder(temp_dir, f"Couldn't delete temporary folder {temp_dir}", True)
 
 
 def create_ssh_keyfile(private_key_file):
@@ -98,22 +99,23 @@ def get_ssh_cert_validity(cert_file):
     return None
 
 
-def write_ssh_config(config_path, resource_group, vm_name, overwrite,
-                     ip, username, cert_file, private_key_file, delete_keys):
+def write_ssh_config(config_path, resource_group, vm_name, overwrite, port,
+                     ip, username, cert_file, private_key_file, delete_keys, delete_cert):
 
-    # Warn users to delete credentials once config file is no longer being used.
-    # If user provided keys, only ask them to delete the certificate.
-    path_to_delete = os.path.dirname(cert_file)
-    items_to_delete = " (id_rsa, id_rsa.pub, id_rsa.pub-aadcert.pub)"
-    if not delete_keys:
-        path_to_delete = cert_file
-        items_to_delete = ""
-    validity = get_ssh_cert_validity(cert_file)
-    validity_warning = ""
-    if validity:
-        validity_warning = f" {validity.lower()}"
-    logger.warning("%s contains sensitive information%s%s\nPlease delete it once you no longer need this config file. ",
-                   path_to_delete, items_to_delete, validity_warning)
+    if delete_keys or delete_cert:
+        # Warn users to delete credentials once config file is no longer being used.
+        # If user provided keys, only ask them to delete the certificate.
+        path_to_delete = os.path.dirname(cert_file)
+        items_to_delete = " (id_rsa, id_rsa.pub, id_rsa.pub-aadcert.pub)"
+        if not delete_keys:
+            path_to_delete = cert_file
+            items_to_delete = ""
+        validity = get_ssh_cert_validity(cert_file)
+        validity_warning = ""
+        if validity:
+            validity_warning = f" {validity.lower()}"
+        logger.warning("%s contains sensitive information%s%s\nPlease delete it once you no longer need this config file. ",
+                    path_to_delete, items_to_delete, validity_warning)
 
     lines = [""]
 
@@ -121,9 +123,12 @@ def write_ssh_config(config_path, resource_group, vm_name, overwrite,
         lines.append("Host " + resource_group + "-" + vm_name)
         lines.append("\tUser " + username)
         lines.append("\tHostName " + ip)
-        lines.append("\tCertificateFile " + cert_file)
+        if cert_file:
+            lines.append("\tCertificateFile " + cert_file)
         if private_key_file:
             lines.append("\tIdentityFile " + private_key_file)
+        if port:
+            lines.append("\tPort " + port)
 
     # default to all hosts for config
     if not ip:
@@ -131,9 +136,12 @@ def write_ssh_config(config_path, resource_group, vm_name, overwrite,
 
     lines.append("Host " + ip)
     lines.append("\tUser " + username)
-    lines.append("\tCertificateFile " + cert_file)
+    if cert_file:
+        lines.append("\tCertificateFile " + cert_file)
     if private_key_file:
         lines.append("\tIdentityFile " + private_key_file)
+    if port:
+        lines.append("\tPort " + port)
 
     if overwrite:
         mode = 'w'
@@ -173,15 +181,17 @@ def _get_host(username, ip):
 def _build_args(cert_file, private_key_file, port):
     private_key = []
     port_arg = []
+    certificate = []
     if private_key_file:
         private_key = ["-i", private_key_file]
     if port:
         port_arg = ["-p", port]
-    certificate = ["-o", "CertificateFile=" + cert_file]
+    if cert_file:
+        certificate = ["-o", "CertificateFile=" + cert_file]
     return private_key + certificate + port_arg
 
 
-def _do_cleanup(delete_keys, cert_file, private_key, log_file=None, wait=False):
+def _do_cleanup(delete_keys, delete_cert, cert_file, private_key, log_file=None, wait=False):
     # if there is a log file, use it to check for the connection success
     if log_file:
         t0 = time.time()
@@ -205,5 +215,5 @@ def _do_cleanup(delete_keys, cert_file, private_key, log_file=None, wait=False):
         public_key = private_key + '.pub'
         file_utils.delete_file(private_key, f"Couldn't delete private key {private_key}. ", True)
         file_utils.delete_file(public_key, f"Couldn't delete public key {public_key}. ", True)
-    if cert_file:
+    if delete_cert and cert_file:
         file_utils.delete_file(cert_file, f"Couldn't delete certificate {cert_file}. ", True)
